GMFPlayerOverlyViewController now based on the Protocol (GMFPlayerOverlayViewControllerProtocol) which defines OverlayViewController behaviour All properties and Methods using GMFPlayerOverlayViewController are now using  UIViewController <GMFPlayerOverlayViewControllerProtocol> All properties and Methods using GMFPlayerOverlayView are now using UIView<GMFPlayerControlsProtocol>. In GMFPlayerViewController the standard GMFPLayerOverlayViewController is created if not set before viewDidLoad.
This allows users of the Framework to create completely custom control overlays. (+1 squashed commit)
Squashed commits:
[3777e34] Found a missed reference and had to put some more optional methods into the GMFPlayerControlsProtocol, and therefor had to insert some responds to selector checks.
Will squash this commit into the rest after possible discussion.

+++ GoogleMediaFramework/GMFPlayerControlsView.m
@@ -196,9 +196,9 @@ static const CGFloat kGMFBarPaddingX = 4;
   NSInteger minutes = (durationSecondsRounded / 60) % 60;
   NSInteger hours = durationSecondsRounded / 3600;
   if (hours) {
-    return [NSString stringWithFormat:@"%d:%02d:%02d", hours, minutes, seconds];
+    return [NSString stringWithFormat:@"%ld:%02ld:%02ld", (long) hours, (long) minutes, (long) seconds];
   } else {
-    return [NSString stringWithFormat:@"%d:%02d", minutes, seconds];
+    return [NSString stringWithFormat:@"%ld:%02ld", (long) minutes, (long) seconds];
   }
 }
